define SpecKey type
Instead of using Strings as keys in instance specs, define a SpecKey
type that is a union of a UUID and a String. This type serializes as a
string, and its `FromStr` and `From<String>` impls will try to parse
strings as UUIDs before storing keys in string format. This allows the
control plane to use UUIDs wherever it makes sense while allowing it
(and other users) to continue to use strings where a UUID component ID
is unavailable or is being used by some other spec element (as might be
the case for a disk device and disk backend).

Index most objects in propolis-server using SpecKeys instead of Strings.
This is mostly just a rote change-the-types, make-it-compile exercise,
but it extends to Crucible backends, which deserve some additional
commentary. In the old ensure API, sending a `DiskRequest` would
register disk components with several different identifiers:

- The disk's name, given by the `DiskRequest`'s `name` field, is used
  - as a key in the VM's `DeviceMap`,
  - as the disk device component's key in the VM's instance spec, and
  - to generate the disk backend component's key in the VM's instance
    spec.
- The disk's ID, given as the `id` in the `Volume`-variant
  `VolumeConstructionRequest` in the `DiskRequest`, is reported by the
  Crucible backend's `get_uuid` function and used
  - to register a metrics producer for each disk, and
  - as the key in the VM's `CrucibleBackendMap`.

Now that all new VMs are started using instance specs, use component
keys for everything:

- Entities in the `DeviceMap` are identified by their corresponding
  components' `SpecKey`s.
- Crucible backends in the `CrucibleBackendMap` are also identified by
  their `SpecKey`s.
- APIs that take a Crucible backend ID on their path now take a String
  and not a UUID. The server converts these to a SpecKey before looking
  up backends in the map.
- When a Crucible backend is created, the machine initializer checks to see
  if its `SpecKey` was a `Uuid`. If so, it will register the backend to
  produce metrics.

The intention is that Omicron will use disk IDs as the keys for its
`CrucibleStorageBackend` components in the specs it generates. It can
then also use these IDs as parameters when requesting snapshots or VCR
replacements. The friendly names users give their Omicron disks no
longer appear in the Propolis API at all.