PHD: use stty to widen the effective terminal for Linux guests (#818)
All of PHD's supported Linux guests' serial console drivers assume by
default that they are writing to an 80-character-wide terminal. This
breaks calls to `run_shell_commands` that are longer than 80 characters,
because guests will insert extra spaces and carriage returns to deal
with terminal wrapping. Suppress this by adding an `stty` command to
these guests' login sequences that tells them the terminal is 10,000
characters wide. Add a framework test to validate this.

The new test also works (with no additional changes) for Windows Server
2022. It won't currently work with Server 2016 or 2019, which use the
80x24 terminal backend, so skip this test on these OSes (and add an
affordance that allows a test to ask the framework for the default guest
OS adapter without actually setting up any VM configuration).

With this change, the boot order tests' `run_long_command` function is
obsolete, so remove it.

Finally, tweak `phd_skip!` so that it can take an arbitrary token
sequence as an argument instead of only taking a string literal. This
allows the use of `format!` in a skip macro.

Tested by running this test with all supported guest OS flavors.