server: make clients parse config TOMLs (#810)
Remove config TOML processing from propolis-server. The only thing the
config TOML is used for in Omicron is specifying a bootrom path (and
possibly a version string); turn this into a propolis-server command-line
argument instead. The instance spec-based ensure API already ignored the
server's config TOML, and the non-spec ensure API is about to be removed
anyway, so this is a good time to make this shift.

To avoid breaking existing config TOML users, move the TOML-to-spec
processing logic into the propolis-config-toml crate, and have it output
propolis-client's generated instance spec types. (This would have been a
layering violation previously, but is OK now that propolis-server no longer
uses this crate.) Teach propolis-cli to use this crate to process TOMLs and
to use the instance spec-based ensure endpoint to create and migrate VMs.

Remove the config TOML from the Propolis zone image and change its command
line invocation. Also change PHD's propolis-server invocations as needed.

Finally, tweak the propolis-server and propolis-cli READMEs to describe the
new state of affairs.