add JSON output format to cpuid-gen (#832)
Add a JSON output format to `cpuid-gen`, which prints output suitable for
inserting into an instance spec that can be fed to `propolis-cli`.

This is useful for ad hoc testing of hypervisor enlightenments.
Enlightenments are exposed through special CPUID leaves in the hypervisor
leaf range (0x4000_0000 to 0x4000_FFFF). To expose enlightenments, Propolis
thus needs to pass the correct CPUID values to bhyve. However, if Propolis
starts setting CPUID values for the hypervisor leaves, it will also need to
set them for the standard and extended leaves. This implies that, to enable
enlightenments, Propolis's instance-ensure API (or propolis-standalone's VM
setup logic) will need a complete set of standard/extended CPUID values to
apply. For `propolis-server`, the easiest way to do this is to use
`propolis-cli new --spec`. The goal of this change is to make it easy to
get a CPUID JSON blob to paste into this command's input spec.

While dorking around with this change, I discovered that the `propolis-cli
new --spec` switch was not properly marked as exclusive with the
`--config-toml` switch. Fix that up as well.

Finally, if `propolis-server` fails to convert an API spec to its internal
instance spec type, and the resulting error has a source, print both the
error and the source. This allows the display of messages like "leaf 7 is
specified both with and without a subleaf" when appropriate.

Tested by booting some ad hoc propolis-server VMs using propolis-cli and
the new output from cpuid-gen.