lib: don't send CPUID settings in vCPU device state payload (#848)
Remove the CPUID portion of the vCPU device state payload that gets sent
during live migration. This payload is not currently constructed
correctly due to a bug, but there's no need to fix it directly: a
migration target's received instance spec will always contain a set of
explicitly-applied CPUID settings from the source, even if the
"original" instance spec that was used to start the VM contained no
explicit CPUID values. This means there's no longer any need to transfer
these settings in the device state phase.

Upgrade PHD's cpuid_boot_test to detect this issue and verify the fix.