lib: emulate Hyper-V enlightenment stack (#849)
Add the beginnings of a Hyper-V enlightenment stack to complement the
bhyve enlightenment stack. This change aims to provide the absolute
minimum functionality needed for a hypervisor to claim to expose a
Hyper-V-compatible guest-hypervisor interface. See the doc comments in
the new Hyper-V module for more details.

Add PHD tests that boot VMs with the Hyper-V enlightenment stack enabled
and (if the guest supports it) that try to query the hypervisor
interface from within the guest.

Add a simple propolis-cli option to enable the Hyper-V stack. Once the
stack has some actual features, this may be better expressed as a table
in the propolis-config-toml crate, but this gives users a quick and easy
way to enable the new interface.