lib: add better management of Hyper-V overlay pages (#851)
Add to the Hyper-V module some supporting types to support overlay
pages. A Hyper-V overlay page is a sort of auxiliary page of guest
memory that "covers up" or temporarily "replaces" a specific physical
page in the guest's physical memory, such that accesses to the overlaid
GPA end up accessing the overlay instead. See the doc comments
throughout overlay.rs for many more details.

This commit does not implement first-class migration support for the
overlay manager; instead, migration targets re-create their overlays
during Hyper-V manager import based on the MSR values they've imported.
This runs the risks that (a) the source and target overlay managers will
disagree about the order in which overlays are to be applied, and (b)
the target will overwrite data that was otherwise written to the overlay
page on the source. These risks are mitigated here by (a) having only
one kind of overlay and (b) the fact that guests should generally not
expect to be able to write to an overlay page. First-class overlay
migration support will address both these issues.