server: allow VCR replacement during block backend start
Modify the state driver's VM startup procedure to allow the driver to
process Crucible volume configuration changes while block backends are
being activated. This fixes a livelock that occurs when starting a VM
with a Crucible VCR that points to an unavailable downstairs: the
unavailable downstairs prevents Crucible activation from proceeding;
Nexus sends a corrected VCR that, if applied, would allow the upstairs
to activate; but the state driver never applies the new VCR because it's
blocked trying to activate using the broken VCR.

Modify the PHD VCR replacement smoke test so that it checks this
behavior. Add an affordance to PHD Crucible disks that allows a test to
specify that the disk's generated VCRs should contain an invalid
downstairs IP. Start a VM with a disk configured this way, then replace
the broken VCR with a corrected VCR and verify that the VM boots
normally.