lib: implement reference TSC enlightenment (#856)
Add a Hyper-V reference time enlightenment to the Hyper-V stack. See the
new tsc module's doc comment for details about how the enlightenment
operates.

Add PHD tests that try to verify both that the enlightenment is present
(by querying the current clocksource) and that it keeps time roughly
correctly (by comparing host and guest time readings). The latter test
also takes readings over a migration to help get some coverage of the
time data phase of live migration (though the migration is not currently
cross-machine, so this is a little less interesting than it might
otherwise be).