server: don't spin out on stop requests while starting
Add a PHD test case for this. This also revealed a bug in the request
queue: if a VM is stuck waiting to activate a Crucible volume, and a
stop request arrives, the VCR replacement that might have fixed it will
be rejected, because the queue rejects replacement requests that arrive
after a stop request is queued. Fix this by keeping the same VCR
replacement request disposition when a stop request is received.