Use production propolis-server flags in PHD CI builds (#878)
We use the dev (e.g. non-release) build of propolis-server in PHD to get
debug assertions, which we want, but we should use the `omicron-build`
feature to be in line with production binaries.

`omicron-build` had also acquired code related to failure injection,
which we *also* want in CI so we can test migration failure codepaths.
That is moved to a new `failure-injection` feature flag with a bit of
ceremony to stop us early if we're ever inadvertently packaging a
propolis-server with `omicron-build` and `failure-injection`.

Fixes #412.

---------

Co-authored-by: Greg Colombo <greg@oxidecomputer.com>