this will probably work, but is rancid
if phd-runner decides to run with parallelism > 1, the tmpdir and
artifact directories, which phd-runner will write to from each
framework, are rewritten to be unique per-framework.

this means the directory layout changes based on available parallelism,
and makes the `phd-run-with-args.sh` script somewhat more gross, and
generally is pretty unfortunate. at this point it's just to demonstrate
that something is possible, but this really deserves the requisite
surgery to share one log dir and one artifact dir across the Frameworks.