Test oximeter metrics end to end (#855)
Other tests in Propolis validate that the Oximeter translation of kstats
for a VM are coherent but do so from stubbed kstat readings. Those tests
are fine (and quick!), but don't cover the risk of a translation error
in Propolis reading from a real system, nor the possibility of reading
kstats incorrectly in the first place.

... good news: test doesn't catch anything interesting! The bug I was
looking at is somewhere else.