client: prepare instance spec types for use in sled-agent API
The next step toward implementing virtual platforms in Omicron is to
move instance spec construction from sled-agent up to Nexus. This
requires instance spec types to appear in sled-agent's API. To allow
this:

- Direct the generated client to add a `schemars::JsonSchema` derive to
  all generated types.
- Add `rename = snake_case` attributes to instance spec enum types that
  sled-agent expects to reuse. Omicron requires all of its API
  specifications to pass the OpenAPI linting rules defined in the
  openapi-lint crate (https://github.com/oxidecomputer/openapi-lint);
  among these is a requirement that enum variants all have snake case
  names.

Tests: cargo test; PHD; verified that it's possible to write the desired
Omicron changes using this change; verified that the Falcon repo
continues to build with its propolis-client dependency pointed to this
commit.