Support guest memory sizes >=1028 GiB
Fixes #907.

This change both produces an E820 table for OVMF to consult for highmem
layout, and clamps the RTC CMOS bytes through which we communicate
highmem size to OVMF. The E820 table supplants the RTC CMOS for OVMF's
purposes, so in theory nothing will check those bytes anymore. Even so,
clamping to them to `ffffff` seems more legible.