Fix `dangerous_implicit_autorefs` warning in `dladm`
This [warning] warns about the creation of implicit references to the
dereference of a raw pointer. I've fixed this by using `ptr::addr_of!`
and `slice::from_raw_parts` to construct the slice into `ma_addr`,
avoiding the creation of temporary references through the raw pointer.
In this case, the suggestions from the lint of just adding an additional
explicit `&` reference as a way of saying "this is okay, because I know
the pointer is non-null and well-aligned" would *probably* have been
fine, but this felt less sketchy.

[warning]: https://doc.rust-lang.org/nightly/nightly-rustc/rustc_lint/autorefs/static.DANGEROUS_IMPLICIT_AUTOREFS.html