server: build Tokio runtime using `oxide-tokio-rt` (#913)
As described in [RFD 579], we are now recommending that the
[`oxide-tokio-rt`] crate be used to construct the Tokio runtime in all
production software that uses Tokio. This crate provides common
functions for constructing Tokio runtimes with the recommended
configurations.

In particular, it currently does the following:

- On illumos, configures the runtime to emit DTrace probes, using
  [`tokio-dtrace`].
- Disables Tokio's LIFO slot optimization. This feature is intended to
  improve message-passing latency, but because tasks in the LIFO slot
  do not currently participate in work-stealing, it can result in
  extreme latency spikes in some cases (see omicron#8334 for a worked
  example).

Thus, this commit changes `propolis-server` to construct its Tokio
runtimes using `oxide-tokio-rt` so that these common recommended
configurations are included. This requires enabling Tokio's unstable
features in the workspace.

[RFD 579]: https://rfd.shared.oxide.computer/rfd/0579
[`oxide-tokio-rt`]: https://github.com/oxidecomputer/oxide-tokio-rt
[`tokio-dtrace`]: https://github.com/oxidecomputer/tokio-dtrace