Handle CPUID subleaves more precisely when using Bhyve defaults (#915)
We would report a few CPUID leaves as not having subleaves at all.
Guests that try to interrogate CPUID leaves according to the APM then
could end up very confused since later "subleaves" would just be
repetitions of the single first leaf we queried from the host.

For leaves F, 10, and 8000_0020, I've chosen to hide the
otherwise-unsupported-anyway PQM and PQE features from PQoS. If a guest
tried to use the PQM/PQE features as-advertised, it should get a #GP
anyway. Presumably #GP from a "supported" feature may not be gracefully
handled. For 8000_0026, I'm less committed to returning zeroes, but it
seems like the least-unreasonable thing to present.

I noticed this from `cpuid` in a Linux guest printing thousands of bogus
0x8000_0020 leaves on dogfood. That's PQoS feature information, and the
APM happens to mention that the appropriate way to query this is to
check individual bits in a bit field (implicitly that there are no more
than 32 subleaves on that function). In that case it was technically a
`cpuid` bug that resulted in it trying all the other 2^32-32 leaves, but
we should be better here.

0x8000_0026 is more likely to matter, but it isn't present on Zen 3
parts.