version checks: be honest about missing control files (#950)
If /dev/viona is missing, propolis-server and propolis-standalone would
complain incorrectly about a missing /dev/vmmctl. Fix that and move the 
"which component failed a version check" information out a bit.

Along the way, add some words on why we cache API version information in
the first place, as this does prevent the most obvious plumbing - add 
the filename when open() fails - from working.