Specialize CPUID 8000_001E more reasonably
If leaf 8000_001E is to be discarded, ensure we clear
`TopologyExtensions` so that guests are not told the now-zeroed leaf is
valid

If leaf 8000_001E is to be kept, specialize it for each vCPU. The code
that was present is mostly fine for this purpose. It is not in the list
of leaves Propolis supports specializing because `propolis-server` and
`propolis-standalone` unconditionally set `has_smt == true` (even for
single-vCPU VMs!), which means a specialized leaf 8000_001E will fault
in EDK2 on boot.

The initial CPUID profile must have `TopologyExtensions` set if guests
are expected to use leaf 8000_001E.

This leaf has spooky action properties with Windows guests. To reiterate
https://github.com/oxidecomputer/omicron/pull/9043/files#r2418399831:
if `PerfCtrExtCore` is set and `TopologyExtensions` is not, Windows
Server 2022 has been seen to enter an infinite loop at boot. If both
bits are set, neither bit is set, or `TopologyExtensions` is set without
`PerfCtrExtCore`, Windows Server 2022 appears to be fine.

This (minorly) changes CPUID profiles that had included leaf 8000_001E.
There is no provision for the (incorrect) CPUID profile that
`propolis-{server,standalone}` could provide before, where the initial
APIC ID was read as zero on all cores while having the
`TopologyExtensions` bit set.