propolis-cli should be able to send TOML-defined CPU profiles (#943)
some plumbing I wrote up to help check the CPU profile from Omicron#8728 across
more OS types. I think shortly after 8728 is sorted it makes sense to move a
lot of the CPUID stuff to a crate shared with Propolis, one reason being we
could bake in known-to-Nexus CPUID profiles in `propolis-cli` as well. Then we
could make `--cpuid-profile milan_v1` Just Work without having to copy around
blobs of hex for well-defined profiles. but, this is sufficient for now.