Plumb sled_model and sled_serial into smbios Type1
This is necessary for a4x2 usage where we want the values in the smbios
to match the values in the platform id of the certificates used for
sprockets.

I'm not sure if this will cause any upstream consumer issues that rely
on the the serial number being equivalent to the VM id, or the product
name being equivalent to "OxVM". If so we'll need to find a way to work
in both scenarios.