distinguish probes across NVMe devices
NVMe commands are unique in a device across queue ID (`qid` in various
parts of Propolis) and command ID (`cid` in various parts of Propolis),
but are ambiguous when looking at an entire VM: two different NVMe
devices will have a "queue 1", and their distinct "queue 1" may also
have "command 1234" submitted at the same time.

Pick a device ID for NVMe controllers, and plumb that through to
distinguish queues similar to how we distinguish queues for block
devices: a `devq_id` construction that makes a device and queue unique
across the VM.

In the NVMe case, a `devq_id` is not uniquely identifying across the
life of a VM, as the queues can be destroyed, recreated, the controller
reset, etc. But absent such an administrative operation, this helps
untangle what's happening across a VM with many disks.

Note that block attachment IDs and NVMe IDs are similar in name, and
probably similar in practice, but are tracked distinctly. This is
similar to the difference between block qeuue IDs and NVMe queue IDs:
the architcture is similar, but the mapping between them is more
arbitrary and may change as an implementation detail of Propolis.