update usdt to 0.6.0
dropshot includes usdt probes and as of 0.16.6 depends on usdt 0.6.0 to
do so. this introduces a few duplicate-ish dependencies that feel a bit
unsightly. in particular, usdt 0.5.0 -> 0.6.0 comes with a few major
version bumps of usdt dependencies so they can't simply be unified to
a common latest version.

bumping Propolis is not sufficient to dedupe these dependencies, as we
get them from Crucible, Oximeter, `qorb`, `slog-dtrace`, and
`tokio-dtrace`, as well as `usdt 0.3.5` from `p4rs`. but getting
Propolis ahead of things means once they're all bumped and bumped here,
we'll lose deps on `goblin 0.8.2`, `scroll 0.12.0`, `dof 0.3.0`, and
probably a few others.